ASoC: tlv320aic32x4: Kernel OOPS while entering DAPM standby mode
During the bootup of the kernel, as soon as the DAPM framework kicks in
it pushes the codec into standy mode.

The existing TVL320AIC32x4 codec driver doesn't prepare the clock in
the probe function.
This leads to an OOPS when the DAPM tries to put it into standy by calling
clk_disable_unprepare()

This patch fixes that problem.

Signed-off-by: Bhargav AK <anur.bhargav@gmail.com>